+++ TavisApi/TrueAchievements/Models/SyncHistory.cs
@@ -3,7 +3,7 @@ using Tavis.Models;
 namespace TavisApi.TrueAchievements.Models;
 
 public class SyncHistory {
-	public int Id { get; set; } = 0;
+	public ulong Id { get; set; } = 0;
 	public DateTime? Start { get; set; }
 	public DateTime? End { get; set; }
 	public int? PlayerCount { get; set; }
